Added ByteReaderTest1 test case.
