+++ IllyumL2T.Core.UnitTests/IllyumL2T.Core.UnitTests.csproj
@@ -50,7 +50,9 @@
     </When>
     <Otherwise>
       <ItemGroup>
-        <Reference Include="Microsoft.VisualStudio.QualityTools.UnitTestFramework" />
+        <Reference Include="Microsoft.VisualStudio.QualityTools.UnitTestFramework">
+          <Private>False</Private>
+        </Reference>
       </ItemGroup>
     </Otherwise>
   </Choose>
@@ -65,6 +67,7 @@
     <Compile Include="PositionalValuesFileParserTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="DelimiterSeparatedValuesFieldsSplitterTests.cs" />
+    <Compile Include="MemoryStreamBinaryTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\IllyumL2T.Core\IllyumL2T.Core.csproj">
